+++ src/main/java/com/prgrms/artzip/exibition/dto/projection/ExhibitionForSimpleQuery.java
@@ -1,6 +1,6 @@
 package com.prgrms.artzip.exibition.dto.projection;
 
-import com.prgrms.artzip.exibition.domain.Period;
+import com.prgrms.artzip.exibition.domain.vo.Period;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -13,5 +13,5 @@ public class ExhibitionForSimpleQuery extends ExhibitionBasicForSimpleQuery {
   private Period period;
   private long likeCount;
   private long reviewCount;
-  
+
 }
